relatives bug solve

+++ server/src/controllers/relative.js
@@ -1,10 +1,10 @@
+const { validationResult } = require("express-validator");
 const Relative = require("../models/relative");
 const sendResponse = require("../utils/sendResponse");
 
 const getRelatives = async (req, res) => {
   try {
-    const relatives = await Relative.find({});
-    console.log(relatives);
+    const relatives = await Relative.find({ userId: req.user.id });
 
     if (relatives.length === 0) {
       return sendResponse("No relatives found", res, 404);
@@ -12,11 +12,16 @@ const getRelatives = async (req, res) => {
 
     res.status(200).json({ relatives });
   } catch (err) {
-    return sendResponse("Server Error", res);
+    return sendResponse(err.message, res);
   }
 };
 
 const addRelative = async (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return sendResponse(errors.array(), res, 400);
+  }
+
   try {
     req.body.userId = req.user.id;
 
@@ -26,7 +31,7 @@ const addRelative = async (req, res) => {
 
     res.json({ newRelative });
   } catch (err) {
-    return sendResponse("Server Error", res);
+    return sendResponse(err.message, res);
   }
 };
 
@@ -44,7 +49,6 @@ const updateRelative = async (req, res) => {
   }
 };
 
-// delete single relative
 const deleteRelative = async (req, res) => {
   try {
     const relative = await Relative.findByIdAndDelete(req.params.id);
@@ -54,11 +58,10 @@ const deleteRelative = async (req, res) => {
     }
     sendResponse("Relative deleted", res, 200);
   } catch (err) {
-    return sendResponse("Server Error", res);
+    return sendResponse(err.message, res);
   }
 };
 
-// get single relative
 const getRelative = async (req, res) => {
   try {
     const relative = await Relative.findById(req.params.id);
@@ -69,20 +72,20 @@ const getRelative = async (req, res) => {
 
     res.json({ relative });
   } catch (err) {
-    return sendResponse("Server Error", res);
+    return sendResponse(err.message, res);
   }
 };
 
 const deleteAllRelative = async (req, res) => {
   try {
-    const relative = await Relative.deleteMany({});
+    const relative = await Relative.deleteMany({ userId: req.user.id });
 
     if (!relative) {
       return sendResponse("Unable to delete relatives", res, 404);
     }
     sendResponse("All Relatives deleted", res, 200);
   } catch (err) {
-    return sendResponse("Server Error", res);
+    return sendResponse(err.message, res);
   }
 };
